Backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
pdf2image
pillow
pytesseract
opencv-python-headless
numpy
